+++ src/main/java/thedarkcolour/exdeorum/compat/ModIds.java
@@ -35,6 +35,7 @@ public class ModIds {
     public static final String ENDERIO = "enderio";
     public static final String BIOMES_O_PLENTY = "biomesoplenty";
     public static final String APPLIED_ENERGISTICS_2 = "ae2";
+    public static final String ARS_NOUVEAU = "ars_nouveau";
     public static final String MODEST_FLINT_OVERHAUL = "modestflintoverhaul";
     public static final String FACTORIUM = "factorium";
 }